mlow/noise: FMA-defeat casts in smplGetEnv (#11)
Wrap the four load-bearing products in smplGetEnv's loop in float32(...) so the
Go compiler cannot fuse a*b + c into a single-rounding FMA; the reference rounds
each multiply separately. No divergence was observed beforehand, but the
protective casts AGENTS.md mandates were absent. gennoise KAT still passes.

Source of truth: https://github.com/oxidezap/whatsapp-rust/blob/41095d4e6ba4610e054e9ede3af1d5e88a83faee/wacore/src/voip/mlow/smpl_gennoise.rs

+++ mlow/noise.go
@@ -118,10 +118,10 @@ func smplGetEnv(exc []float32, length int, smthCoef float32, smthState *float32,
 	gainSmthCoef := gainCoef * smthCoef
 	i := 0
 	for i+3 < length {
-		tmp0 := exc[i]*exc[i] + exc[i+1]*exc[i+1]
-		tmp1 := exc[i+2]*exc[i+2] + exc[i+3]*exc[i+3]
-		y1 := gainCoef*tmp1 + gainSmthCoef*tmp0 + smthCoef2*state
-		y0 := gainCoef*tmp0 + smthCoef*state
+		tmp0 := float32(exc[i]*exc[i]) + float32(exc[i+1]*exc[i+1])
+		tmp1 := float32(exc[i+2]*exc[i+2]) + float32(exc[i+3]*exc[i+3])
+		y1 := float32(gainCoef*tmp1) + float32(gainSmthCoef*tmp0) + float32(smthCoef2*state)
+		y0 := float32(gainCoef*tmp0) + float32(smthCoef*state)
 		env[i] = float32(math.Sqrt(float64(y0)))
 		env[i+1] = env[i]
 		env[i+2] = float32(math.Sqrt(float64(y1)))
